Progress
Did the individual page.
Add property page can only be accessed by a agent and the agent can send a request to the admin to add a new property,
Admin approval page added and the agent's property requests goes through, admin can approve or reject. When approved the property gets added to the properties table and page.